fix: store decimal points, parse nested averages, stop swallowing sync errors
This league scores in decimals. The live payload reports a season total as
12.2 and a single match as 4.1, so `points INTEGER` rejected the real data:
the daily sync died on `invalid input syntax for type integer: "12.2"` and
wrote zero player snapshots, which is the entire point of the job.

Rounding was the wrong fix. Expected points are real points here, so
truncating every player to a whole number would bias the exact quantity the
lineup picker optimises. Both points columns become NUMERIC(7, 2).

The failure was invisible because the two writes at the end of syncDaily were
the only unwrapped ones in the job, and /api/cron/sync did not guard syncAll
either, so any throw became a 500 with an empty body and no clue which job
broke. Both now degrade to a warning, matching every other job.

Separately, `average` is an object in this payload -- {average, homeAverage,
awayAverage, averageLastFive, matches, fitness} -- so reading it with num()
silently produced undefined and the column was never populated.

The getNextRound smoke check assumed its fixture was the only future round,
which stopped being true once a real calendar was stored. It now asserts the
contract: the earliest round still ahead.

+++ src/lib/db/repo.ts
@@ -196,7 +196,7 @@ export async function writeSnapshots(
        clause_price, owner_team_id, locked, market_price, captured_at
      )
      SELECT $1::date, * FROM UNNEST(
-       $2::text[], $3::bigint[], $4::int[], $5::numeric[],
+       $2::text[], $3::bigint[], $4::numeric[], $5::numeric[],
        $6::bigint[], $7::text[], $8::boolean[], $9::bigint[]
      ), now()
      ON CONFLICT (snapshot_date, player_id) DO UPDATE SET
@@ -568,7 +568,7 @@ export async function upsertRoundPoints(lineup: RoundLineup): Promise<number> {
        assists, yellow_cards, red_cards, started
      )
      SELECT $1::text, * FROM UNNEST(
-       $2::text[], $3::text[], $4::int[], $5::int[], $6::int[],
+       $2::text[], $3::text[], $4::numeric[], $5::int[], $6::int[],
        $7::int[], $8::int[], $9::int[], $10::boolean[]
      )
      ON CONFLICT (round_id, player_id, team_id) DO UPDATE SET
